backend/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
bcrypt==4.0.0
beanie==1.11.9
bson==0.5.10
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
h11==0.13.0
httptools==0.4.0
idna==3.3
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
pymongo==4.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarl==1.8.1
